+++ src/tests/test-fixtures/basic-expected.scala
@@ -1,8 +1,8 @@
 trait Person {
   val name: String
-  val age: Int
+  val age: Double
 }
 
 type Color = "red" | "green" | "blue"
 
-type Age = Int | String
+type Age = Double | String
